Update a version of qa-cqs. It can support execution on the IonQ simulator and hardware.
