+++ cqs_module/expansion.py
@@ -27,28 +27,22 @@
     Another is our self-developed method: Ansatz tree with l1 sampling expansion.
 """
 
-from cqs_module.calculator import U_list_dagger
+from cqs_module.calculation import U_list_dagger, calculate_statistics
 
-# from hardware.Qibo.qibo_access import Hadamard_test
-# from hardware.IonQ.ionq_access import Hadamard_test as Hadamard_test_ibmq
-from hardware.Eigens.eigens_access import Hadamard_test as Hadamard_test_eigens
-from hardware.IBMQ.ibmq_access import Hadamard_test as Hadamard_test_ibmq
+from hardware.execute import Hadamard_test
 
-from numpy import random, linalg, sqrt, kron, array
-from numpy import real, imag, conj, transpose
+from numpy import random, linalg, sqrt, kron
+from numpy import real, conj, transpose
 from utils import draw_ansatz_tree
-
-from cqs_module.calculator import verify_gradient_overlap, get_x, get_unitary, zero_state
-from cqs_module.mitigation import Richardson_extrapolate
-
+from qiskit_ionq import IonQProvider
 
 # unitaries = [['X', 'Z'], ['Z', 'X'], ['I', 'Y']]
 # alphas = [1.6317321653264614, 0.2249104575899552, 1.7897631234464821]
 
 
-def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_budget=1024, frugal=False):
-    if mtd is None:
-        mtd = 'Hadamard'
+def expand_ansatz_tree(A, vars, ansatz_tree, backend=None, draw_tree=False, shots_budget=1024, frugal=False):
+    if backend is None:
+        backend = 'eigens'
     A_coeffs = A.get_coeff()
     A_unitaries = A.get_unitary()
     A_terms_number = len(A_coeffs)
@@ -59,39 +53,94 @@ def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_bu
         draw_ansatz_tree(A_terms_number, tree_depth, 'Expansion')
     child_space = [parent_node + A_unitaries[i] for i in range(A_terms_number)]
 
-    if mtd == 'Hadamard':
-        if frugal is True:
-            shots_each_entry = shots_budget / (10 * len(child_space))
+    if frugal is True:
+        shots_each_entry = shots_budget / (10 * len(child_space))
 
-            M_Child_Nodes = sum([abs(A_coeffs[k] * A_coeffs[l] * vars[j])
-                                 for j in range(tree_depth)
-                                 for k in range(A_terms_number)
-                                 for l in range(A_terms_number)] +
-                                [abs(A_coeffs[j]) for j in range(A_terms_number)])
+        M_Child_Nodes = sum([abs(A_coeffs[k] * A_coeffs[l] * vars[j])
+                             for j in range(tree_depth)
+                             for k in range(A_terms_number)
+                             for l in range(A_terms_number)] +
+                            [abs(A_coeffs[j]) for j in range(A_terms_number)])
 
-            P_Child_Nodes_Term_1 = [10 * int(shots_each_entry * abs(vars[j] * A_coeffs[k] * A_coeffs[l]) / M_Child_Nodes)
-                                    for j in range(tree_depth)
-                                    for k in range(A_terms_number)
-                                    for l in range(A_terms_number)]
-            P_Child_Nodes_Term_2 = [10 * int(shots_each_entry * abs(A_coeffs[j]) / M_Child_Nodes) for j in
-                                    range(A_terms_number)]
+        P_Child_Nodes_Term_1 = [10 * int(shots_each_entry * abs(vars[j] * A_coeffs[k] * A_coeffs[l]) / M_Child_Nodes)
+                                for j in range(tree_depth)
+                                for k in range(A_terms_number)
+                                for l in range(A_terms_number)]
+        P_Child_Nodes_Term_2 = [10 * int(shots_each_entry * abs(A_coeffs[j]) / M_Child_Nodes) for j in
+                                range(A_terms_number)]
+
+    else:
+        # Uniform distribution
+        shots_ave = 100
+        P_Child_Nodes_Term_1 = [shots_ave
+                                for _ in range(tree_depth)
+                                for _ in range(A_terms_number)
+                                for _ in range(A_terms_number)]
+        P_Child_Nodes_Term_2 = [shots_ave for j in range(A_terms_number)]
+
+    gradient_overlaps = [0 for _ in range(len(child_space))]
+
+    for i, child_node in enumerate(child_space):
+        if backend == 'ionq':
+            Job_ids_1_R = []
+            Job_ids_1_I = []
+            Job_ids_2_R = []
+            Job_ids_2_I = []
+            for j in range(tree_depth):
+                anstaz_state = ansatz_tree[j]
+                for k in range(A_terms_number):
+                    for l in range(A_terms_number):
+                        shots = P_Child_Nodes_Term_1[j * A_terms_number * A_terms_number + k * A_terms_number + l]
+                        u = U_list_dagger(child_node) + A_unitaries[k] + A_unitaries[l] + anstaz_state
+                        # shots = 100
+                        jobid_R = Hadamard_test(u, backend=backend, alpha=1, shots=shots)
+                        Job_ids_1_R.append(jobid_R)
+                        jobid_I = Hadamard_test(u, backend=backend, alpha=1j, shots=shots)
+                        Job_ids_1_I.append(jobid_I)
+
+            for j in range(A_terms_number):
+                shots = P_Child_Nodes_Term_2[j]
+                u = U_list_dagger(child_node) + A_unitaries[j]
+                # shots = 100
+                jobid_R = Hadamard_test(u, backend=backend, alpha=1, shots=shots)
+                Job_ids_2_R.append(jobid_R)
+                jobid_I = Hadamard_test(u, backend=backend, alpha=1j, shots=shots)
+                Job_ids_2_I.append(jobid_I)
+
+            provider = IonQProvider('pUhwyKCHRYAvWUChFqwTApQwow4mS2h7')
+            # simulator_backend = provider.get_backend("ionq_qpu.harmony")
+            # simulator_backend = provider.get_backend("ionq_qpu.aria-1")
+            simulator_backend = provider.get_backend("ionq_simulator")
+
+            exp_1_R = calculate_statistics(simulator_backend.retrieve_jobs(Job_ids_1_R))
+            exp_1_I = calculate_statistics(simulator_backend.retrieve_jobs(Job_ids_1_I))
+            exp_2_R = calculate_statistics(simulator_backend.retrieve_jobs(Job_ids_2_R))
+            exp_2_I = calculate_statistics(simulator_backend.retrieve_jobs(Job_ids_2_I))
+
+            term_1 = 0
+            for j in range(tree_depth):
+                term_1_1 = 0
+                alpha = vars[j]
+                anstaz_state = ansatz_tree[j]
+                for k in range(A_terms_number):
+                    for l in range(A_terms_number):
+                        beta_k = A_coeffs[k]
+                        beta_l = A_coeffs[l]
+                        inner_product_real = exp_1_R[j * A_terms_number * A_terms_number + k * A_terms_number + l]
+                        inner_product_imag = exp_1_I[j * A_terms_number * A_terms_number + k * A_terms_number + l]
+                        inner_product = inner_product_real - inner_product_imag * 1j
+                        term_1_1 += beta_k * beta_l * inner_product
+                term_1 += alpha * term_1_1
+
+            term_2 = 0
+            for j in range(A_terms_number):
+                beta_j = A_coeffs[j]
+                inner_product_real = exp_2_R[j]
+                inner_product_imag = exp_2_R[j]
+                inner_product = inner_product_real - inner_product_imag * 1j
+                term_2 += beta_j * inner_product
 
         else:
-            # Uniform distribution
-            # shots_ave = int(shots_budget / (len(child_space) * (tree_depth * (A_terms_number ** 2) + A_terms_number)))
-            shots_ave = 12
-            P_Child_Nodes_Term_1 = [shots_ave
-                                    for _ in range(tree_depth)
-                                    for _ in range(A_terms_number)
-                                    for _ in range(A_terms_number)]
-            P_Child_Nodes_Term_2 = [shots_ave for j in range(A_terms_number)]
-            # print("Number of shots for gradient expansion:", shots)
-            # print()
-
-
-        gradient_overlaps = [0 for _ in range(len(child_space))]
-
-        for i, child_node in enumerate(child_space):
             term_1 = 0
             for j in range(tree_depth):
                 term_1_1 = 0
@@ -102,9 +151,10 @@ def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_bu
                         beta_k = A_coeffs[k]
                         beta_l = A_coeffs[l]
                         shots = P_Child_Nodes_Term_1[j * A_terms_number * A_terms_number + k * A_terms_number + l]
+                        # shots = 100
                         u = U_list_dagger(child_node) + A_unitaries[k] + A_unitaries[l] + anstaz_state
-                        inner_product_real = Hadamard_test_ibmq(u, shots=shots)
-                        inner_product_imag = Hadamard_test_ibmq(u, alpha=1j, shots=shots)
+                        inner_product_real = Hadamard_test(u, backend=backend, alpha=1, shots=shots)
+                        inner_product_imag = Hadamard_test(u, backend=backend, alpha=1j, shots=shots)
                         inner_product = inner_product_real - inner_product_imag * 1j
                         term_1_1 += beta_k * beta_l * inner_product
                 term_1 += alpha * term_1_1
@@ -113,96 +163,27 @@ def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_bu
             for j in range(A_terms_number):
                 beta_j = A_coeffs[j]
                 shots = P_Child_Nodes_Term_2[j]
+                # shots = 100
                 u = U_list_dagger(child_node) + A_unitaries[j]
-                inner_product_real = Hadamard_test_ibmq(u, shots=shots)
-                inner_product_imag = Hadamard_test_ibmq(u, alpha=1j, shots=shots)
+                inner_product_real = Hadamard_test(u, backend=backend, alpha=1, shots=shots)
+                inner_product_imag = Hadamard_test(u, backend=backend, alpha=1j, shots=shots)
                 inner_product = inner_product_real - inner_product_imag * 1j
                 term_2 += beta_j * inner_product
 
-            gradient_overlap = abs(2 * term_1 - 2 * term_2)
+        gradient_overlap = abs(2 * term_1 - 2 * term_2)
+        gradient_overlaps[i] = gradient_overlap
 
-            gradient_overlaps[i] = gradient_overlap
-
-        # To consider the case when there are several candidates for the child node.
-        max_index = [index for index, item in enumerate(gradient_overlaps) if item == max(gradient_overlaps)]
-
-        # verify_gradient_overlap(A, vars, ansatz_tree, max_index)
-
-        # candidate_states = [child_space[i] for i in max_index]
-        # print("Next child node can be chosen from:", candidate_states)
-        idx = random.choice(max_index)
-        child_node_opt = child_space[idx]
-    # print("Next child node is selected as:", child_node_opt)
-
-    # Use the matrix calculations to decide the child nodes
-    elif mtd == 'Matrix':
-        A_mat = A.get_matrix()
-        x = get_x(vars, ansatz_tree)
-        zeros = zero_state()
-        width = len(A_unitaries[0][0])
-        if width > 1:
-            for j in range(width - 1):
-                zeros = kron(zeros, zero_state())
-
-        parent_node = ansatz_tree[-1]
-        child_space = [parent_node + A_unitaries[i] for i in range(A_terms_number)]
-        gradient_overlaps = [0 for _ in range(len(child_space))]
-
-        for i, child_node in enumerate(child_space):
-            U_mat = get_unitary(child_node)
-
-            gradient = 2 * conj(transpose(zeros)) @ conj(transpose(U_mat)) @ A_mat @ A_mat @ x - 2 * conj(
-                transpose(zeros)) @ conj(transpose(U_mat)) @ A_mat @ zeros
-            gradient_overlaps[i] = abs(gradient.item())
-
-        # print('Matrix Calculation:', gradient_overlaps)
-
-        max_index = [index for index, item in enumerate(gradient_overlaps) if item == max(gradient_overlaps)]
-        idx = random.choice(max_index)
-        child_node_opt = child_space[idx]
-
-    elif mtd == 'Eigens':
-        gradient_overlaps = [0 for _ in range(len(child_space))]
-        for i, child_node in enumerate(child_space):
-            term_1 = 0
-            for j in range(tree_depth):
-                term_1_1 = 0
-                alpha = vars[j]
-                anstaz_state = ansatz_tree[j]
-                for k in range(A_terms_number):
-                    for l in range(A_terms_number):
-                        beta_k = A_coeffs[k]
-                        beta_l = A_coeffs[l]
-                        u = U_list_dagger(child_node) + A_unitaries[k] + A_unitaries[l] + anstaz_state
-                        inner_product = Hadamard_test_eigens(u)
-                        term_1_1 += beta_k * beta_l * inner_product
-                term_1 += alpha * term_1_1
-
-            term_2 = 0
-            for j in range(A_terms_number):
-                beta_j = A_coeffs[j]
-                u = U_list_dagger(child_node) + A_unitaries[j]
-                inner_product = Hadamard_test_eigens(u)
-                term_2 += beta_j * inner_product
-
-            gradient_overlaps[i] = abs(2 * term_1 - 2 * term_2)
-
-        # print("Child nodes are:", child_space)
-        # print('Estimated Gradient overlaps are:', gradient_overlaps)
-
-        # To consider the case when there are several candidates for the child node.
-        max_index = [index for index, item in enumerate(gradient_overlaps) if item == max(gradient_overlaps)]
-
-        # verify_gradient_overlap(A, vars, ansatz_tree, max_index)
-
-        # candidate_states = [child_space[i] for i in max_index]
-        # print("Next child node can be chosen from:", candidate_states)
-        idx = random.choice(max_index)
-        child_node_opt = child_space[idx]
-    # print("Next child node is selected as:", child_node_opt)
-    else:
-        raise ValueError
-
+    print("")
+    print("Child space is:", child_space)
+    print("")
+    print("Gradient Overlaps are:", gradient_overlaps)
+    print("")
+    # To consider the case when there are several candidates for the child node.
+    max_index = [index for index, item in enumerate(gradient_overlaps) if item == max(gradient_overlaps)]
+    idx = random.choice(max_index)
+    child_node_opt = child_space[idx]
+    print("Choose the child node of maximum overlap:", child_node_opt)
+    print("")
     if draw_tree is True:
         draw_ansatz_tree(A_terms_number, tree_depth, 'Optimal', which_index=idx)
     ansatz_tree.append(child_node_opt)
@@ -211,65 +192,65 @@ def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_bu
     return ansatz_tree
 
 
-def optimize_with_stochastic_descend(A, N):
-    # Use L1 sampling to do the expansion.
-    # c.f. https://arxiv.org/abs/1907.05378
-    A_coeffs = A.get_coeff()
-    A_unitaries = A.get_unitary()
-
-    width = len(A_unitaries[0][0])
-    zeros = zero_state()
-    if width > 1:
-        for j in range(width - 1):
-            zeros = kron(zeros, zero_state())
-
-    # xt = array([0 for _ in range(2 ** width)]).reshape(-1, 1)
-    xt = zeros
-    XT = [xt]
-
-    # print("B is", B)
-    # print('eta is:', eta)
-    # print('Upper bound is:', error)
-
-    A_coeffs_norm = [coeff / sum(A_coeffs) for coeff in A_coeffs]
-    unitary_indexes = list(range(len(A_unitaries)))
-
-    A_mat = A.get_matrix()
-
-    # In order to quantify the L2, we will have to pre-know the optimal x by matrix calculations
-    x_opt = linalg.inv(A_mat) @ zeros
-
-    for t in range(1, N):
-        L2 = linalg.norm(xt - x_opt)
-        B = 4 * (L2 ** 2) + 4
-        eta = L2 / (B * sqrt(N))
-        # print(L2)
-        U1, U2, U3 = [A_unitaries[random.choice(unitary_indexes, p=A_coeffs_norm)] for _ in range(3)]
-        gt = 2 * get_unitary(U1) @ get_unitary(U2) @ xt - 2 * get_unitary(U3) @ zeros
-        xt = xt - eta * gt
-        # print('eta is:', eta)
-        XT.append(xt)
-
-    # print(XT)
-    # print()
-    # print(sum(XT))
-    # print()
-    xt_ave = sum(XT) / N
-    # print(xt_ave)
-
-    loss = real((conj(transpose(xt_ave)) @ conj(transpose(A_mat)) @ A_mat @ xt_ave - 2 * real(
-        conj(transpose(zeros)) @ A_mat @ xt_ave)).item()) + 1
-    # print("Loss:", loss)
-
-    return loss
-
-
-
-
-
-    # size = decomposition_terms ** depth
-
-    # current_step = child_space[0]
+# def optimize_with_stochastic_descend(A, N):
+#     # Use L1 sampling to do the expansion.
+#     # c.f. https://arxiv.org/abs/1907.05378
+#     A_coeffs = A.get_coeff()
+#     A_unitaries = A.get_unitary()
+#
+#     width = len(A_unitaries[0][0])
+#     zeros = zero_state()
+#     if width > 1:
+#         for j in range(width - 1):
+#             zeros = kron(zeros, zero_state())
+#
+#     # xt = array([0 for _ in range(2 ** width)]).reshape(-1, 1)
+#     xt = zeros
+#     XT = [xt]
+#
+#     # print("B is", B)
+#     # print('eta is:', eta)
+#     # print('Upper bound is:', error)
+#
+#     A_coeffs_norm = [coeff / sum(A_coeffs) for coeff in A_coeffs]
+#     unitary_indexes = list(range(len(A_unitaries)))
+#
+#     A_mat = A.get_matrix()
+#
+#     # In order to quantify the L2, we will have to pre-know the optimal x by matrix calculations
+#     x_opt = linalg.inv(A_mat) @ zeros
+#
+#     for t in range(1, N):
+#         L2 = linalg.norm(xt - x_opt)
+#         B = 4 * (L2 ** 2) + 4
+#         eta = L2 / (B * sqrt(N))
+#         # print(L2)
+#         U1, U2, U3 = [A_unitaries[random.choice(unitary_indexes, p=A_coeffs_norm)] for _ in range(3)]
+#         gt = 2 * get_unitary(U1) @ get_unitary(U2) @ xt - 2 * get_unitary(U3) @ zeros
+#         xt = xt - eta * gt
+#         # print('eta is:', eta)
+#         XT.append(xt)
+#
+#     # print(XT)
+#     # print()
+#     # print(sum(XT))
+#     # print()
+#     xt_ave = sum(XT) / N
+#     # print(xt_ave)
+#
+#     loss = real((conj(transpose(xt_ave)) @ conj(transpose(A_mat)) @ A_mat @ xt_ave - 2 * real(
+#         conj(transpose(zeros)) @ A_mat @ xt_ave)).item()) + 1
+#     # print("Loss:", loss)
+#
+#     return loss
+#
+#
+#
+#
+#
+#     # size = decomposition_terms ** depth
+#
+#     # current_step = child_space[0]
 
 
 
@@ -284,9 +265,9 @@ def optimize_with_stochastic_descend(A, N):
 
 
 ########################################################################################################################
-# def expand_ansatz_tree(A, vars, ansatz_tree, mtd=None, draw_tree=False, shots_power=4):
-#     if mtd is None:
-#         mtd = 'Hadamard'
+# def expand_ansatz_tree(A, vars, ansatz_tree, backend=None, draw_tree=False, shots_power=4):
+#     if backend is None:
+#         backend = 'Hadamard'
 #     A_coeffs = A.get_coeff()
 #     A_unitaries = A.get_unitary()
 #
@@ -303,7 +284,7 @@ def optimize_with_stochastic_descend(A, N):
 #         draw_ansatz_tree(A_terms_number, tree_depth, 'Expansion')
 #     child_space = [parent_node + A_unitaries[i] for i in range(A_terms_number)]
 #
-#     if mtd == 'Hadamard':
+#     if backend == 'Hadamard':
 #         gradient_overlaps = [0 for _ in range(len(child_space))]
 #
 #         for i, child_node in enumerate(child_space):
@@ -416,7 +397,7 @@ def optimize_with_stochastic_descend(A, N):
 #     # print("Next child node is selected as:", child_node_opt)
 #
 #     # Use the matrix calculations to decide the child nodes
-#     elif mtd == 'Matrix':
+#     elif backend == 'Matrix':
 #         A_mat = A.get_matrix()
 #         x = get_x(vars, ansatz_tree)
 #         zeros = zero_state()
@@ -442,7 +423,7 @@ def optimize_with_stochastic_descend(A, N):
 #         idx = random.choice(max_index)
 #         child_node_opt = child_space[idx]
 #
-#     elif mtd == 'Eigens':
+#     elif backend == 'Eigens':
 #         gradient_overlaps = [0 for _ in range(len(child_space))]
 #         for i, child_node in enumerate(child_space):
 #             term_1 = 0
